Restructuring sqlite3 and makefiles.
sqlite3 is moved from third_party/ to third_party/sqlite3/ and the rest
of the files from SQLite3 amalgamation is added.
Remove build directory and build in place (object files created in same
dir as source files).
Update makefiles.
All tests passes.

+++ test/program.cpp
@@ -36,39 +36,39 @@ int main(int argc, char** argv) {
     std::cout << sim_db.read<bool>("new_param4") << std::endl;
 
     std::vector<int> param5 = sim_db.read<std::vector<int> >("param5");
-    for (int i = 0; i < param5.size(); i++) {
+    for (size_t i = 0; i < param5.size(); i++) {
         std::cout << param5[i] << std::endl;
     }
     sim_db.write<std::vector<int> >("new_param5", param5);
-    for (int i = 0; i < param5.size(); i++) {
+    for (size_t i = 0; i < param5.size(); i++) {
         std::cout << param5[i] << std::endl;
     }
 
     std::vector<double> param6 = sim_db.read<std::vector<double> >("param6");
-    for (int i = 0; i < param6.size(); i++) {
+    for (size_t i = 0; i < param6.size(); i++) {
         std::cout << param6[i] << std::endl;
     }
     sim_db.write<std::vector<double> >("new_param6", param6);
-    for (int i = 0; i < param6.size(); i++) {
+    for (size_t i = 0; i < param6.size(); i++) {
         std::cout << param6[i] << std::endl;
     }
 
     std::vector<std::string> param7 =
             sim_db.read<std::vector<std::string> >("param7");
-    for (int i = 0; i < param7.size(); i++) {
+    for (size_t i = 0; i < param7.size(); i++) {
         std::cout << param7[i] << std::endl;
     }
     sim_db.write<std::vector<std::string> >("new_param7", param7);
-    for (int i = 0; i < param7.size(); i++) {
+    for (size_t i = 0; i < param7.size(); i++) {
         std::cout << param7[i] << std::endl;
     }
 
     std::vector<bool> param8 = sim_db.read<std::vector<bool> >("param8");
-    for (int i = 0; i < param8.size(); i++) {
+    for (size_t i = 0; i < param8.size(); i++) {
         std::cout << param8[i] << std::endl;
     }
     sim_db.write<std::vector<bool> >("new_param8", param8);
-    for (int i = 0; i < param8.size(); i++) {
+    for (size_t i = 0; i < param8.size(); i++) {
         std::cout << param8[i] << std::endl;
     }
 
@@ -86,7 +86,7 @@ int main(int argc, char** argv) {
     // Save param6 to file in this unique subdirectory.
     std::ofstream result_file;
     result_file.open(filename_result);
-    for (int i = 0; i < param6.size(); i++) {
+    for (size_t i = 0; i < param6.size(); i++) {
         result_file << param6[i] << std::endl;
     }
 }
